refactor(web): extract PublicShareRow into its own file (ROK-1067)
LineupDetailHeader.tsx grew to 320 lines after the public-share work,
crossing the 300-line ESLint cap. Move PublicShareRow + the
PublicShareCopyOnly sibling into LineupPublicShareRow.tsx — leaves the
header at 273 lines and keeps the public-share concern in a single
testable module.
